新增供应商退货管理模块 修复库存材料型号null查询失败bug 招商合同新增endflag终止标识，新增终止合同协议相关功能

+++ src/main/java/com/yuesheng/pm/restapi/ZujinApi.java
@@ -543,8 +543,10 @@ public class ZujinApi extends BaseApi {
     @Operation(description = "更新租金合同")
     @PostMapping
     public ResponseModel update(@RequestBody ProZujin zujin, @SessionAttribute(Constant.SESSION_KEY) Staff staff) {
-        houseService.deleteByZujin(zujin.getId());
-        houseService.insertHouseRelation(zujin.getHouses(), zujin);
+        if (zujin.getHouses() != null){
+            houseService.deleteByZujin(zujin.getId());
+            houseService.insertHouseRelation(zujin.getHouses(), zujin);
+        }
         zujin.setLastStaffId(staff.getId());
         zujin.setLastDateTime(DateUtil.format(new Date()));
         return new ResponseModel(zujinService.update(zujin));
